Split datagrid tests into files
