+++ tests/ids-data-grid/ids-data-grid-paging.spec.ts
@@ -1,0 +1,246 @@
+import { expect } from '@playwright/test';
+import { test } from '../base-fixture';
+
+import IdsDataGrid from '../../src/components/ids-data-grid/ids-data-grid';
+import IdsPagerInput from '../../src/components/ids-pager/ids-pager-input';
+
+test.describe('IdsDataGrid paging tests', () => {
+  test.describe('client-side paging tests', () => {
+    const clientPagingUrl = '/ids-data-grid/pagination-client-side.html';
+
+    test.beforeEach(async ({ page }) => {
+      await page.goto(clientPagingUrl);
+    });
+
+    test('renders pager', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      const dataGridEl = await page.locator('ids-data-grid');
+      await expect(await dataGridEl.getAttribute('pagination')).toEqual('client-side');
+
+      const pagerInputEl = await page.locator('ids-data-grid ids-pager-input');
+      await expect(await pagerInputEl?.getAttribute('page-number')).toEqual('1');
+    });
+
+    test('clear data', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      const pagerNextBtn = await page.locator('ids-pager-button[next]');
+      await pagerNextBtn.click();
+
+      await (await page.locator('ids-data-grid [aria-rowindex="5"] [aria-colindex="1"]')).click();
+
+      const titleText = await page.locator('#title-text');
+      await expect(await titleText.textContent()).toEqual('1 Result');
+
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('2');
+
+      const pagerPrevBtn = await page.locator('ids-pager-button[previous]');
+      await pagerPrevBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('1');
+
+      const clearBtn = await page.locator('[aria-label="Clear Row"]');
+      await clearBtn.click();
+
+      await expect(await titleText.textContent()).toEqual('');
+
+      await pagerNextBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('2');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="5"] [aria-colindex="3"]').textContent()).toEqual('');
+
+      const pagerLastBtn = await page.locator('ids-pager-button[last]');
+      await pagerLastBtn.click();
+
+      await (await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="1"]')).click();
+
+      const pagerFirstBtn = await page.locator('ids-pager-button[first]');
+      await pagerFirstBtn.click();
+
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('1');
+      await clearBtn.click();
+
+      await pagerLastBtn.click();
+      await expect(await titleText.textContent()).toEqual('');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="3"]').textContent()).toEqual('');
+    });
+
+    test('remove data', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      const pagerNextBtn = await page.locator('ids-pager-button[next]');
+      await pagerNextBtn.click();
+
+      await (await page.locator('ids-data-grid [aria-rowindex="5"] [aria-colindex="1"]')).click();
+
+      const titleText = await page.locator('#title-text');
+      await expect(await titleText.textContent()).toEqual('1 Result');
+
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('2');
+
+      const pagerPrevBtn = await page.locator('ids-pager-button[previous]');
+      await pagerPrevBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('1');
+
+      const deleteBtn = await page.locator('[aria-label="Delete Row"]');
+      await deleteBtn.click();
+
+      await expect(await titleText.textContent()).toEqual('');
+
+      await pagerNextBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('2');
+
+      await expect(await page.locator('ids-data-grid [aria-rowindex="10"] [aria-colindex="2"]').textContent()).toEqual('21');
+
+      const pagerLastBtn = await page.locator('ids-pager-button[last]');
+      await pagerLastBtn.click();
+
+      await (await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="1"]')).click();
+
+      const pagerFirstBtn = await page.locator('ids-pager-button[first]');
+      await pagerFirstBtn.click();
+
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('1');
+      await deleteBtn.click();
+
+      await pagerLastBtn.click();
+      await expect(await titleText.textContent()).toEqual('');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('993');
+    });
+
+    test('navigates pages', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      const pagerInputEl = await page.locator('ids-data-grid ids-pager-input');
+      await expect(await pagerInputEl?.getAttribute('page-number')).toEqual('1');
+
+      // Set number input
+      const pagerInputHandle = await page.$('ids-data-grid ids-pager-input');
+      await pagerInputHandle?.evaluate((el: IdsPagerInput) => {
+        el.setAttribute('page-number', '2');
+      });
+      await expect(await pagerInputHandle?.getAttribute('page-number')).toEqual('2');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('11');
+
+      // Click next button
+      const pagerNextBtn = await page.locator('ids-pager-button[next]');
+      await pagerNextBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('3');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('21');
+
+      // Click previous button
+      const pagerPrevBtn = await page.locator('ids-pager-button[previous]');
+      await pagerPrevBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('2');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('11');
+
+      // Click first button
+      const pagerFirstBtn = await page.locator('ids-pager-button[first]');
+      await pagerFirstBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('1');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('1');
+
+      // Click last button
+      const pagerLastBtn = await page.locator('ids-pager-button[last]');
+      await pagerLastBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('100');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('991');
+    });
+
+    test('selects across pages', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      // Check two items on first page
+      await (await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="1"]')).click();
+      await (await page.locator('ids-data-grid [aria-rowindex="3"] [aria-colindex="1"]')).click();
+
+      // Click next button
+      const pagerNextBtn = await page.locator('ids-pager-button[next]');
+      await pagerNextBtn.click();
+
+      // Check two items on second page
+      await (await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="1"]')).click();
+      await (await page.locator('ids-data-grid [aria-rowindex="3"] [aria-colindex="1"]')).click();
+
+      const handle = await page.$('ids-data-grid');
+      const selectedRows = await handle?.evaluate((el: IdsDataGrid) => el?.selectedRowsAcrossPages);
+      await expect(selectedRows?.length).toEqual(4);
+    });
+  });
+
+  test.describe('server-side paging tests', () => {
+    const clientPagingUrl = '/ids-data-grid/pagination-server-side.html';
+
+    test.beforeEach(async ({ page }) => {
+      await page.goto(clientPagingUrl);
+    });
+
+    test('renders pager', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      const dataGridEl = await page.locator('ids-data-grid');
+      await expect(await dataGridEl.getAttribute('pagination')).toEqual('server-side');
+
+      const pagerInputEl = await page.locator('ids-data-grid ids-pager-input');
+      await expect(await pagerInputEl?.getAttribute('page-number')).toEqual('1');
+    });
+
+    test('navigates pages', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      const pagerInputEl = await page.locator('ids-data-grid ids-pager-input');
+      await expect(await pagerInputEl?.getAttribute('page-number')).toEqual('1');
+
+      // Set number input
+      const pagerInputHandle = await page.$('ids-data-grid ids-pager-input');
+      await pagerInputHandle?.evaluate((el: IdsPagerInput) => {
+        el.setAttribute('page-number', '2');
+      });
+      await expect(await pagerInputHandle?.getAttribute('page-number')).toEqual('2');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('11');
+
+      // Click next button
+      const pagerNextBtn = await page.locator('ids-pager-button[next]');
+      await pagerNextBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('3');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('21');
+
+      // Click previous button
+      const pagerPrevBtn = await page.locator('ids-pager-button[previous]');
+      await pagerPrevBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('2');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('11');
+
+      // Click first button
+      const pagerFirstBtn = await page.locator('ids-pager-button[first]');
+      await pagerFirstBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('1');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('1');
+
+      // Click last button
+      const pagerLastBtn = await page.locator('ids-pager-button[last]');
+      await pagerLastBtn.click();
+      await expect(await page.locator('ids-data-grid ids-pager-input').getAttribute('page-number')).toEqual('100');
+      await expect(await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="2"]').textContent()).toEqual('991');
+    });
+
+    test('selects across pages', async ({ page, browserName }) => {
+      if (browserName !== 'chromium') return;
+
+      // Check two items on first page
+      await (await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="1"]')).click();
+      await (await page.locator('ids-data-grid [aria-rowindex="3"] [aria-colindex="1"]')).click();
+
+      // Click next button
+      const pagerNextBtn = await page.locator('ids-pager-button[next]');
+      await pagerNextBtn.click();
+
+      // Check two items on second page
+      await (await page.locator('ids-data-grid [aria-rowindex="1"] [aria-colindex="1"]')).click();
+      await (await page.locator('ids-data-grid [aria-rowindex="3"] [aria-colindex="1"]')).click();
+
+      const handle = await page.$('ids-data-grid');
+      const selectedRows = await handle?.evaluate((el: IdsDataGrid) => el?.selectedRowsAcrossPages);
+      await expect(selectedRows?.length).toEqual(4);
+    });
+  });
+});
